+++ example/lib/main.dart
@@ -40,7 +40,7 @@ class _MyHomePageState extends State<MyHomePage> with SingleTickerProviderStateM
 
   final List<String> _list = [
     '0','SDk','plugin updates','Facebook','哔了狗了QP又不够了',
-    'Kirchhoff','Italy','France','Spain','Dart','Foo','Select','lorem ip','9',
+    'Kirchhoff','Italy','France','Spain','美','Dart','Foo','Select','lorem ip','9',
     'Star','Flutter Selectable Tags','1','Hubble','2','Input flutter tags','A B C','8','Android Studio developer','welcome to the jungle','Gauss',
       '美术',
       '互联网',
@@ -97,6 +97,7 @@ class _MyHomePageState extends State<MyHomePage> with SingleTickerProviderStateM
             '美术',
             'substring',
             'last tag',
+            '术',
             'enable',
             'act',
             '1',
@@ -236,10 +237,13 @@ class _MyHomePageState extends State<MyHomePage> with SingleTickerProviderStateM
                               fontSize: _fontSize,
                               symmetry: _symmetry,
                               singleItem: _singleItem,
-                              //offset: 25,
+                              //offset: 0, // to be used in combination with the padding
                               //activeColor: Colors.deepPurple,
                               //boxShadow: [],
-                              //margin: EdgeInsets.symmetric(horizontal: 3, vertical: 6),
+                              //borderRadius:5,
+                              //margin: EdgeInsets.symmetric(horizontal: 50, vertical: 6),
+                              //padding: EdgeInsets.symmetric(horizontal: 18),
+                              //height: 28,
                               onPressed: (tag){
                                 setState(() {
                                   _selectableOnPressed = tag.toString();
@@ -337,18 +341,17 @@ class _MyHomePageState extends State<MyHomePage> with SingleTickerProviderStateM
                                   symmetry: _symmetry,
                                   iconBackground: Colors.green[800],
                                   //boxShadow: [],
+                                  //offset: 5,
                                   //color: Colors.red,
-                                  //margin: EdgeInsets.all(15),
-                                  //iconPadding: EdgeInsets.all(10),
-                                  //iconMargin: EdgeInsets.all(0),
+                                  //padding: EdgeInsets.only(left: 10),
+                                  //margin: EdgeInsets.symmetric(horizontal: 20, vertical: 6),
+                                  //iconPadding: EdgeInsets.all(5),
+                                  //iconMargin: EdgeInsets.only(right:5,left: 2),
                                   lowerCase: true,
-                                  autofocus: false,
-                                  onDelete: (tag){
-                                      print(tag);
-                                  },
-                                  onInsert: (tag){
-                                      print(tag);
-                                  },
+                                  autofocus: true,
+                                  onDelete: (tag) => print(tag),
+                                  onInsert: (tag) => print(tag),
+
                               ),
                           ),
                           Padding(
@@ -385,7 +388,7 @@ class _MyHomePageState extends State<MyHomePage> with SingleTickerProviderStateM
   {
     List<DropdownMenuItem> list = [];
 
-    int count = 15;
+    int count = 19;
 
     for(int i = 1; i < count; i++)
       list.add(